share utils
